refactor project-tools add-block boundaries
