+++ .sampo/changesets/issue-347-cli-add-block-boundaries.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split `cli-add-block` config generation and legacy validator repair into focused runtime helpers while preserving existing workspace add-block behavior.
